+++ lib/dependency-analysis/getNodeModuleDependencies.ts
@@ -1,37 +1,13 @@
 import * as ts from "typescript"
 import * as path from "path"
 import * as fs from "fs"
-
-interface NodeModuleDependency {
-  /** The package name (e.g., "react" or "@tscircuit/core") */
-  packageName: string
-  /** The import path used in the source (e.g., "react/jsx-runtime") */
-  importPath: string
-  /** Resolved absolute file paths for this dependency */
-  resolvedFiles: string[]
-}
-
-function getAllDependencyPackages(projectDir: string): Set<string> {
-  const packageJsonPath = path.join(projectDir, "package.json")
-  const allPackages = new Set<string>()
-
-  if (!fs.existsSync(packageJsonPath)) {
-    return allPackages
-  }
-
-  try {
-    const packageJson = JSON.parse(fs.readFileSync(packageJsonPath, "utf-8"))
-    const deps = packageJson.dependencies || {}
-
-    for (const packageName of Object.keys(deps)) {
-      allPackages.add(packageName)
-    }
-  } catch (error) {
-    console.warn("Failed to parse package.json for dependencies:", error)
-  }
-
-  return allPackages
-}
+import { isRuntimeProvidedPackage } from "../utils/isRuntimeProvidedPackage"
+import { getPackageNameFromImport } from "../utils/getPackageNameFromImport"
+import { getPackageNameFromFilePath } from "../utils/getPackageNameFromFilePath"
+import { getAllDependencyPackages } from "../utils/getAllDependencyPackages"
+import { findPackageDir } from "../utils/findPackageDir"
+import { findPackageDirFromResolvedFile } from "../utils/findPackageDirFromResolvedFile"
+import { collectPackageFiles } from "../utils/collectPackageFiles"
 
 /**
  * Extracts all node_modules imports from a source file
@@ -104,68 +80,8 @@ export function getNodeModuleImports(filePath: string): string[] {
 }
 
 /**
- * Extracts the package name from an import path
- * e.g., "react/jsx-runtime" -> "react"
- * e.g., "@tscircuit/core/components" -> "@tscircuit/core"
- */
-function getPackageNameFromImport(importPath: string): string {
-  if (importPath.startsWith("@")) {
-    // Scoped package
-    const parts = importPath.split("/")
-    return `${parts[0]}/${parts[1]}`
-  }
-  // Regular package
-  return importPath.split("/")[0]
-}
-
-/**
- * Finds the package directory in node_modules for a given package name
- */
-function findPackageDir({
-  packageName,
-  projectDir,
-  searchFromDir,
-}: {
-  packageName: string
-  projectDir: string
-  searchFromDir?: string
-}): string | undefined {
-  // Try to find the package in multiple locations:
-  // 1. Project's node_modules
-  // 2. If searchFromDir is provided (e.g., inside a local package), search upwards
-  const searchPaths: string[] = [
-    path.join(projectDir, "node_modules", packageName),
-  ]
-
-  if (searchFromDir) {
-    // Walk up the directory tree from searchFromDir to find node_modules
-    let currentDir = path.dirname(searchFromDir)
-    const projectDirNormalized = path.normalize(projectDir)
-
-    while (currentDir.startsWith(projectDirNormalized)) {
-      const candidatePath = path.join(currentDir, "node_modules", packageName)
-      if (!searchPaths.includes(candidatePath)) {
-        searchPaths.push(candidatePath)
-      }
-
-      const parentDir = path.dirname(currentDir)
-      if (parentDir === currentDir) break
-      currentDir = parentDir
-    }
-  }
-
-  for (const candidatePath of searchPaths) {
-    if (fs.existsSync(candidatePath)) {
-      return candidatePath
-    }
-  }
-
-  return undefined
-}
-
-/**
- * Resolves a node_modules import to actual file paths
- * Searches in both the project's node_modules and parent directories for hoisted packages
+ * Resolves a node_modules import to actual file paths.
+ * Searches in both the project's node_modules and parent directories for hoisted packages.
  */
 export function resolveNodeModuleImport({
   importPath,
@@ -177,7 +93,6 @@ export function resolveNodeModuleImport({
   searchFromDir?: string
 }): string[] {
   const packageName = getPackageNameFromImport(importPath)
-
   const packageDir = findPackageDir({ packageName, projectDir, searchFromDir })
 
   if (!packageDir) {
@@ -186,7 +101,7 @@ export function resolveNodeModuleImport({
 
   const resolvedFiles: string[] = []
 
-  // If importing a subpath (e.g., "react/jsx-runtime" or "kicad-libraries/footprints/...")
+  // If importing a subpath (e.g., "react/jsx-runtime")
   if (importPath !== packageName) {
     const subpath = importPath.slice(packageName.length + 1)
     const possiblePaths = [
@@ -208,7 +123,6 @@ export function resolveNodeModuleImport({
       }
     }
 
-    // If we found the subpath file, return it (we have what we need)
     if (resolvedFiles.length > 0) {
       return resolvedFiles
     }
@@ -220,7 +134,6 @@ export function resolveNodeModuleImport({
     try {
       const packageJson = JSON.parse(fs.readFileSync(packageJsonPath, "utf-8"))
 
-      // Main entry points from package.json
       const entryPoints = [
         packageJson.main,
         packageJson.module,
@@ -240,6 +153,7 @@ export function resolveNodeModuleImport({
     }
   }
 
+  // Fallback to common entry point locations
   if (resolvedFiles.length === 0) {
     const fallbackPaths = [
       path.join(packageDir, "index.js"),
@@ -265,7 +179,7 @@ export function resolveNodeModuleImport({
 }
 
 /**
- * Recursively collects all node_modules files used by a source file and its dependencies
+ * Recursively collects all node_modules dependencies from a source file and its local imports
  */
 export function collectAllNodeModuleDependencies(
   entryFilePath: string,
@@ -273,7 +187,7 @@ export function collectAllNodeModuleDependencies(
   maxDepth = 10,
 ): Map<string, string[]> {
   const visited = new Set<string>()
-  const nodeModuleFiles = new Map<string, string[]>() // importPath -> resolved files
+  const nodeModuleFiles = new Map<string, string[]>()
 
   function processFile(filePath: string, depth: number) {
     if (depth > maxDepth || visited.has(filePath)) {
@@ -382,177 +296,20 @@ export function collectAllNodeModuleDependencies(
 }
 
 /**
- * Directories that should be excluded when collecting package files
+ * Gets all file paths from node_modules that should be uploaded for a project.
+ * Only uploads packages explicitly listed in dependencies (excluding runtime-provided packages).
  */
-const EXCLUDED_PACKAGE_DIRECTORIES = new Set([
-  "node_modules",
-  ".git",
-  ".next",
-  ".turbo",
-  "coverage",
-  ".cache",
-  "tmp",
-  "temp",
-])
-
-/**
- * Recursively collects all files from a package directory
- * Prioritizes transpiled files (dist/, build/) over source files to avoid
- * path alias resolution issues in the browser
- */
-function collectPackageFiles(packageDir: string): string[] {
-  // First, check if there's a dist/ or build/ directory with transpiled files
-  const buildDirs = ["dist", "build"]
-
-  // Check build directories in order
-  for (const dirName of buildDirs) {
-    const dirPath = path.join(packageDir, dirName)
-    if (fs.existsSync(dirPath)) {
-      const files = walkDirectory(dirPath, new Set())
-      if (files.length > 0) {
-        // Also include package.json for metadata if it exists
-        const packageJsonPath = path.join(packageDir, "package.json")
-        if (fs.existsSync(packageJsonPath)) {
-          files.push(packageJsonPath)
-        }
-        return files
-      }
-    }
-  }
-
-  // Fall back to collecting all source files (excluding build directories)
-  return walkDirectory(packageDir, EXCLUDED_PACKAGE_DIRECTORIES)
-}
-
-function walkDirectory(dir: string, excludedDirs: Set<string>): string[] {
-  const files: string[] = []
-
-  if (!fs.existsSync(dir)) return files
-
-  const entries = fs.readdirSync(dir, { withFileTypes: true })
-
-  for (const entry of entries) {
-    const fullPath = path.join(dir, entry.name)
-
-    if (entry.isDirectory()) {
-      // Skip excluded directories
-      if (excludedDirs.has(entry.name)) {
-        continue
-      }
-      files.push(...walkDirectory(fullPath, excludedDirs))
-    } else if (entry.isFile()) {
-      files.push(fullPath)
-    }
-  }
-
-  return files
-}
-
-const RUNTIME_PROVIDED_PACKAGES = new Set([
-  "react",
-  "react-dom",
-  "react/jsx-runtime",
-  "tscircuit",
-  "@tscircuit/core",
-  "@tscircuit/props",
-])
-
-/**
- * Check if a package is provided by the runtime and should not be uploaded
- */
-function isRuntimeProvidedPackage(packageName: string): boolean {
-  // Check exact match
-  if (RUNTIME_PROVIDED_PACKAGES.has(packageName)) {
-    return true
-  }
-
-  // Check if it's a subpath of a runtime package
-  for (const runtimePkg of Array.from(RUNTIME_PROVIDED_PACKAGES)) {
-    if (packageName.startsWith(`${runtimePkg}/`)) {
-      return true
-    }
-  }
-
-  return false
-}
-
-/**
- * Check if a file path is inside a runtime-provided package's node_modules directory
- * This handles cases where runtime packages are incorrectly bundled as transitive dependencies
- * e.g., node_modules/@tsci/some-package/node_modules/react/jsx-runtime.js
- */
-function isFileInRuntimeProvidedPackage(filePath: string): boolean {
-  const normalizedPath = path.normalize(filePath)
-  const pathSegments = normalizedPath.split(path.sep)
-
-  // Find all node_modules occurrences and check the package after each
-  for (let i = 0; i < pathSegments.length - 1; i++) {
-    if (pathSegments[i] === "node_modules") {
-      // Get the package name after node_modules
-      const nextSegment = pathSegments[i + 1]
-      if (!nextSegment) continue
-
-      let packageName: string
-      if (nextSegment.startsWith("@") && pathSegments[i + 2]) {
-        // Scoped package: @scope/package
-        packageName = `${nextSegment}/${pathSegments[i + 2]}`
-      } else {
-        packageName = nextSegment
-      }
-
-      if (isRuntimeProvidedPackage(packageName)) {
-        return true
-      }
-    }
-  }
-
-  return false
-}
-
-/**
- * Finds the package directory from a resolved file path
- */
-function findPackageDirFromResolvedFile(
-  resolvedFile: string,
-  packageName: string,
-): string | undefined {
-  // Walk up from the resolved file to find the package directory
-  let packageDir = path.dirname(resolvedFile)
-
-  while (packageDir.includes("node_modules")) {
-    // Check if this directory has a package.json with matching name
-    const packageJsonPath = path.join(packageDir, "package.json")
-    if (fs.existsSync(packageJsonPath)) {
-      try {
-        const pkgJson = JSON.parse(fs.readFileSync(packageJsonPath, "utf-8"))
-        if (pkgJson.name === packageName) {
-          return packageDir
-        }
-      } catch {
-        // Continue searching
-      }
-    }
-
-    const nextParentDir = path.dirname(packageDir)
-    if (nextParentDir === packageDir) break
-    packageDir = nextParentDir
-  }
-
-  return undefined
-}
-
 export function getAllNodeModuleFilePaths(
   entryFilePath: string,
   projectDir: string,
 ): string[] {
   const allDependencyPackages = getAllDependencyPackages(projectDir)
 
-  // Early return if no dependencies are defined
   if (allDependencyPackages.size === 0) {
     return []
   }
 
-  // Collect all node_modules dependencies from the entry file
+  // Collect dependencies from imports in the entry file
   const dependencies = collectAllNodeModuleDependencies(
     entryFilePath,
     projectDir,
@@ -576,38 +333,26 @@ export function getAllNodeModuleFilePaths(
   const processedPackages = new Set<string>()
   const allFiles = new Set<string>()
 
-  // Iterate through all discovered dependencies
   for (const [importPath, resolvedFiles] of Array.from(
     dependencies.entries(),
   )) {
     const packageName = getPackageNameFromImport(importPath)
-
-    // Check if this package is in the project's dependencies
     const isProjectDependency = allDependencyPackages.has(packageName)
 
-    // Skip runtime-provided packages
-    if (isRuntimeProvidedPackage(packageName)) {
+    // Skip runtime-provided packages and non-project dependencies
+    if (isRuntimeProvidedPackage(packageName) || !isProjectDependency) {
       continue
     }
 
-    // Only upload packages that are explicitly listed in dependencies
-    if (!isProjectDependency) {
-      continue
-    }
-
-    // Upload project dependencies
     if (!processedPackages.has(packageName)) {
       processedPackages.add(packageName)
 
-      // Use the first resolved file to find the package directory
       if (resolvedFiles.length > 0) {
-        const firstResolvedFile = resolvedFiles[0]
         const packageDir = findPackageDirFromResolvedFile(
-          firstResolvedFile,
+          resolvedFiles[0],
           packageName,
         )
 
-        // Collect all files from the package directory
         if (packageDir && fs.existsSync(packageDir)) {
           const packageFiles = collectPackageFiles(packageDir)
           packageFiles.forEach((file) => allFiles.add(file))
@@ -616,11 +361,9 @@ export function getAllNodeModuleFilePaths(
     }
   }
 
-  // Filter out any files that are inside runtime-provided packages
-  // This handles cases where packages incorrectly bundle react/tscircuit as dependencies
-  const filteredFiles = Array.from(allFiles).filter(
-    (file) => !isFileInRuntimeProvidedPackage(file),
-  )
-
-  return filteredFiles
+  // Filter out files inside runtime-provided packages (handles transitive deps)
+  return Array.from(allFiles).filter((file) => {
+    const pkgName = getPackageNameFromFilePath(file)
+    return !pkgName || !isRuntimeProvidedPackage(pkgName)
+  })
 }
